Init Account Trigger
Write a trigger that creates two identical Opportunities whenever an Account is created. Make sure both Opportunities are associated with the Account. Use any values for the fields on the Opportunities - just make sure to use variables when populating the fields of each Opportunity to make sure they are identical.